refactor hotel rating packaging
