+++ src/main/java/com/amadeus/Amadeus.java
@@ -48,11 +48,19 @@ public class Amadeus extends HTTPClient {
    */
   public Shopping shopping;
 
+  /**
+   * <p>
+   * A namespaced client for the <code>/v2/e-reputation</code> endpoints.
+   * </p>
+   */
+  public EReputation ereputation;
+
   protected Amadeus(Configuration configuration) {
     super(configuration);
     this.referenceData = new ReferenceData(this);
     this.travel = new Travel(this);
     this.shopping = new Shopping(this);
+    this.ereputation = new EReputation(this);
   }
 
   /**
