+++ src/main/java/com/amadeus/Shopping.java
@@ -7,7 +7,6 @@ import com.amadeus.shopping.FlightOffersSearch;
 import com.amadeus.shopping.HotelOffer;
 import com.amadeus.shopping.HotelOffers;
 import com.amadeus.shopping.HotelOffersByHotel;
-import com.amadeus.shopping.HotelRatings;
 
 /**
  * <p>
@@ -77,14 +76,6 @@ public class Shopping {
   public HotelOffersByHotel hotelOffersByHotel;
 
   /**
-   * <p>
-   *   A namespaced client for the
-   *   <code>/v2/e-reputation/hotel-sentiments</code> endpoints.
-   * </p>
-   */
-  public HotelRatings hotelRatings;
-
-  /**
    * Constructor.
    * @hide
    */
@@ -96,8 +87,6 @@ public class Shopping {
     this.hotelOffers = new HotelOffers(client);
     this.hotelOffersByHotel = new HotelOffersByHotel(client);
     this.flightOffersSearch = new FlightOffersSearch(client);
-    this.hotelRatings = new HotelRatings(client);
-
   }
 
   /**
